+++ webapp/tests/integration/plugin-api.test.js
@@ -3,8 +3,8 @@
  * @tagline         Integration tests for W-045 Plugin API authentication
  * @description     Tests plugin REST API authentication and authorization
  * @file            webapp/tests/integration/plugin-api.test.js
- * @version         1.3.3
- * @release         2025-12-01
+ * @version         1.3.4
+ * @release         2025-12-02
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/