+++ webapp/tests/unit/utils/cache-manager.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for CacheManager utility
  * @description     Tests for centralized cache management with periodic refresh
  * @file            webapp/tests/unit/utils/cache-manager.test.js
- * @version         1.3.3
- * @release         2025-12-01
+ * @version         1.3.4
+ * @release         2025-12-02
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/